+++ src/main/java/com/braintreegateway/test/MerchantAccount.java
@@ -2,4 +2,9 @@ package com.braintreegateway.test;
 
 public class MerchantAccount {
     public static String Approve = "approve_me";
+
+    public static String InsufficientFundsContactUs = "insufficient_funds__contact";
+    public static String AccountNotAuthorizedContactUs = "account_not_authorized__contact";
+    public static String BankRejectedUpdateFundingInformation = "bank_rejected__update";
+    public static String BankRejectedNone = "bank_rejected__none";
 }