+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.29.0
+
+* Include Dispute information on Transaction
+* Search for Transactions disputed on a certain date
+
+## 2.28.0
+
+* Disbursement Webhooks
+
 ## 2.27.0
 * Merchant account find API